+++ landscape/managers/tools/QRTool.h
@@ -15,7 +15,7 @@
 
 struct ToolData {
     ToolData() = default;
-    ToolData(QRMatrix<double> *hmap, size_t w, size_t h,
+    ToolData(QRMatrix<float> *hmap, size_t w, size_t h,
             double wS, const QRVectorIterator<sptr<QRFrame2D>> &plates,
             const QRVectorIterator<Vector3D> &move): hmap(hmap), width(w), height(h), worldStep(wS),
             plates(plates), moveVectors(move) {
@@ -25,7 +25,7 @@ struct ToolData {
     QRVector<sptr<QRFrame2D>> plates;
     QRVector<Vector3D> moveVectors;
 
-    QRMatrix<double> *hmap = nullptr;
+    QRMatrix<float> *hmap = nullptr;
     size_t width, height;     // point width&height (points count, not polygons)
     double worldStep;  // step - polyStep (1,2,4 etc), x-y-Step - coordinate steps
 };
